clean up interfaces and rename dtos
