--- src/users/interfaces/user.interface.ts
@@ -1,5 +1,0 @@
-export interface User {
-  id: number;
-  name: string;
-  age: number;
-}
