+++ src/users/users.service.ts
@@ -1,18 +1,17 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { UserEntity } from './entity/user.entity';
-import { User } from './interfaces/user.interface';
+import { User } from './entity/user.entity';
 
 @Injectable()
 export class UsersService {
   constructor(
-    @InjectRepository(UserEntity)
-    private usersRepository: Repository<UserEntity>,
+    @InjectRepository(User)
+    private usersRepository: Repository<User>,
   ) {}
 
   async create(name: string, age: number): Promise<void> {
-    const user = new UserEntity();
+    const user = new User();
     user.name = name;
     user.age = age;
     await this.usersRepository.save(user);
@@ -25,7 +24,7 @@ export class UsersService {
     await this.usersRepository.update({ id }, user);
   }
 
-  async find(id: number): Promise<User | null> {
+  async find(id: number) {
     return this.usersRepository.findOneBy({ id });
   }
